pyproject.toml
[tool.poetry]
name = "cms"
version = "0.1.0"
description = "CMS API for managing audio and video content"
authors = ["Aghyad"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = { extras = ["standard"], version = "^0.35.0" }
pydantic = {extras = ["email"], version = "^2.11.7"}
SQLAlchemy = "^2.0.43"
asyncpg = "^0.30.0"
alembic = "^1.16.4"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.1.1"
pydantic-settings = "^2.10.1"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
celery = "^5.5.3"
redis = "^6.4.0"
opensearch-py = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
httpx = "^0.27.0"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
